+++ freshservice/table_announcement.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"fmt"
 	fs "github.com/theapsgroup/go-freshservice/freshservice"
-	"github.com/turbot/steampipe-plugin-sdk/v4/grpc/proto"
-	"github.com/turbot/steampipe-plugin-sdk/v4/plugin"
+	"github.com/turbot/steampipe-plugin-sdk/v5/grpc/proto"
+	"github.com/turbot/steampipe-plugin-sdk/v5/plugin"
 )
 
 func tableAnnouncement() *plugin.Table {
@@ -116,7 +116,7 @@ func announcementColumns() []*plugin.Column {
 
 // Hydrate Functions
 func getAnnouncement(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
-	id := int(d.KeyColumnQuals["id"].GetInt64Value())
+	id := int(d.EqualsQuals["id"].GetInt64Value())
 
 	client, err := connect(ctx, d)
 	if err != nil {
@@ -154,7 +154,7 @@ func listAnnouncements(ctx context.Context, d *plugin.QueryData, h *plugin.Hydra
 		}
 	}
 
-	q := d.KeyColumnQuals
+	q := d.EqualsQuals
 	if q["state"] != nil {
 		filter.State = q["state"].GetStringValue()
 	}