+++ freshservice/table_problem_task.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"fmt"
 	fs "github.com/theapsgroup/go-freshservice/freshservice"
-	"github.com/turbot/steampipe-plugin-sdk/v4/grpc/proto"
-	"github.com/turbot/steampipe-plugin-sdk/v4/plugin"
-	"github.com/turbot/steampipe-plugin-sdk/v4/plugin/transform"
+	"github.com/turbot/steampipe-plugin-sdk/v5/grpc/proto"
+	"github.com/turbot/steampipe-plugin-sdk/v5/plugin"
+	"github.com/turbot/steampipe-plugin-sdk/v5/plugin/transform"
 )
 
 func tableProblemTask() *plugin.Table {
@@ -100,7 +100,7 @@ func problemTaskColumns() []*plugin.Column {
 
 // Hydrate Functions
 func listProblemTasks(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
-	problemId := int(d.KeyColumnQuals["problem_id"].GetInt64Value())
+	problemId := int(d.EqualsQuals["problem_id"].GetInt64Value())
 
 	client, err := connect(ctx, d)
 	if err != nil {