+++ freshservice/table_solution_article.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"fmt"
 	fs "github.com/theapsgroup/go-freshservice/freshservice"
-	"github.com/turbot/steampipe-plugin-sdk/v4/grpc/proto"
-	"github.com/turbot/steampipe-plugin-sdk/v4/plugin"
+	"github.com/turbot/steampipe-plugin-sdk/v5/grpc/proto"
+	"github.com/turbot/steampipe-plugin-sdk/v5/plugin"
 )
 
 func tableSolutionArticle() *plugin.Table {
@@ -131,7 +131,7 @@ func solutionArticleColumns() []*plugin.Column {
 
 // Hydrate Functions
 func getSolutionArticle(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
-	id := int(d.KeyColumnQuals["id"].GetInt64Value())
+	id := int(d.EqualsQuals["id"].GetInt64Value())
 
 	client, err := connect(ctx, d)
 	if err != nil {
@@ -169,7 +169,7 @@ func listSolutionArticles(ctx context.Context, d *plugin.QueryData, h *plugin.Hy
 		}
 	}
 
-	q := d.KeyColumnQuals
+	q := d.EqualsQuals
 	if q["folder_id"] != nil {
 		fid := int(q["folder_id"].GetInt64Value())
 		filter.FolderID = fid