+++ freshservice/table_ticket_timeentry.go
@@ -3,9 +3,9 @@ package freshservice
 import (
 	"context"
 	"fmt"
-	"github.com/turbot/steampipe-plugin-sdk/v4/grpc/proto"
-	"github.com/turbot/steampipe-plugin-sdk/v4/plugin"
-	"github.com/turbot/steampipe-plugin-sdk/v4/plugin/transform"
+	"github.com/turbot/steampipe-plugin-sdk/v5/grpc/proto"
+	"github.com/turbot/steampipe-plugin-sdk/v5/plugin"
+	"github.com/turbot/steampipe-plugin-sdk/v5/plugin/transform"
 )
 
 func tableTicketTimeEntry() *plugin.Table {
@@ -93,7 +93,7 @@ func ticketTimeEntryColumns() []*plugin.Column {
 
 // Hydrate Functions
 func listTicketTimeEntries(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
-	ticketId := int(d.KeyColumnQuals["ticket_id"].GetInt64Value())
+	ticketId := int(d.EqualsQuals["ticket_id"].GetInt64Value())
 
 	client, err := connect(ctx, d)
 	if err != nil {